+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Feature: Adding acm certificate for ssl
 - fix: CKV_AWS_92  #Ensure the ELB has access logging enabled. For the minimal example, this is not enabled
+- fix: `out-of-service` status for registered targets
 
 ## [1.3.0] - 2023-10-05
 - feat: removed s3 module resource from the root module and showcased it in complete example.